requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
cryptography==41.0.7

# Token Sync Dependencies
apscheduler==3.10.4
httpx==0.25.0